made github helper hook function and passed npm test Co-authored-by: mbuenrostro21 <mbuenrostro21@users.noreply.github.com> Co-authored-by: Daniel Park <parkdaniel731@users.noreply.github.com> Co-authored-by: fpena213 <fpena213@users.noreply.github.com> Co-authored-by: Michelle Conroy <missmshel@users.noreply.github.com>
