+++ src/lib/components/LoginBox.svelte
@@ -1,6 +1,5 @@
 <a class="github" href="/oauth/github/api"
-    >Sign in with Github NPM version jfnbaskjfbnjkasfjk</a
+    >Sign in with Github NPM version</a
 >
-
 <style>
 </style>
